+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/materialization/PredicateSplit.java
@@ -2,8 +2,12 @@
 
 package com.starrocks.sql.optimizer.rule.transformation.materialization;
 
+import com.google.common.collect.Lists;
 import com.starrocks.sql.optimizer.Utils;
+import com.starrocks.sql.optimizer.operator.scalar.BinaryPredicateOperator;
 import com.starrocks.sql.optimizer.operator.scalar.ScalarOperator;
+
+import java.util.List;
 
 public class PredicateSplit {
     // column equality predicates conjuncts
@@ -44,7 +48,40 @@ public class PredicateSplit {
         return residualPredicates;
     }
 
-    public ScalarOperator toScalarOperator() {
-        return Utils.compoundAnd(equalPredicates, rangePredicates, residualPredicates);
+    // split predicate into three parts: equal columns predicates, range predicates, and residual predicates
+    public static PredicateSplit splitPredicate(ScalarOperator predicate) {
+        if (predicate == null) {
+            return PredicateSplit.of(null, null, null);
+        }
+        List<ScalarOperator> predicateConjuncts = Utils.extractConjuncts(predicate);
+        List<ScalarOperator> columnEqualityPredicates = Lists.newArrayList();
+        List<ScalarOperator> rangePredicates = Lists.newArrayList();
+        List<ScalarOperator> residualPredicates = Lists.newArrayList();
+        for (ScalarOperator scalarOperator : predicateConjuncts) {
+            if (scalarOperator instanceof BinaryPredicateOperator) {
+                BinaryPredicateOperator binary = (BinaryPredicateOperator) scalarOperator;
+                ScalarOperator leftChild = scalarOperator.getChild(0);
+                ScalarOperator rightChild = scalarOperator.getChild(1);
+                if (binary.getBinaryType().isEqual()) {
+                    if (leftChild.isColumnRef() && rightChild.isColumnRef()) {
+                        columnEqualityPredicates.add(scalarOperator);
+                    } else if (leftChild.isColumnRef() && rightChild.isConstantRef()) {
+                        rangePredicates.add(scalarOperator);
+                    } else {
+                        residualPredicates.add(scalarOperator);
+                    }
+                } else if (binary.getBinaryType().isRange()) {
+                    if (leftChild.isColumnRef() && rightChild.isConstantRef()) {
+                        rangePredicates.add(scalarOperator);
+                    } else {
+                        residualPredicates.add(scalarOperator);
+                    }
+                }
+            } else {
+                residualPredicates.add(scalarOperator);
+            }
+        }
+        return PredicateSplit.of(Utils.compoundAnd(columnEqualityPredicates), Utils.compoundAnd(rangePredicates),
+                Utils.compoundAnd(residualPredicates));
     }
 }
